refactor: extract node filtering to a separate file
