+++ internal/route/routing_strategy.go
@@ -6,9 +6,6 @@ import (
 	"sort"
 	"sync/atomic"
 
-	"github.com/satsuma-data/node-gateway/internal/checks"
-	"github.com/satsuma-data/node-gateway/internal/metadata"
-
 	"go.uber.org/zap"
 	"golang.org/x/exp/maps"
 )
@@ -47,23 +44,4 @@ func (s *PriorityRoundRobinStrategy) RouteNextRequest(upstreamsByPriority map[in
 	}
 
 	return "", ErrNoHealthyUpstreams
-}
-
-type RequestMetadata struct{}
-
-type NodeFilter interface {
-	Apply(requestMetadata *RequestMetadata, upstreamID string) bool
-}
-
-type IsHealthyAndAtMaxHeightFilter struct {
-	healthCheckManager checks.HealthCheckManager
-	chainMetadataStore *metadata.ChainMetadataStore
-}
-
-func (f *IsHealthyAndAtMaxHeightFilter) Apply(_ *RequestMetadata, upstreamID string) bool {
-	var maxHeight = f.chainMetadataStore.GetMaxHeight()
-
-	var upstreamStatus = f.healthCheckManager.GetUpstreamStatus(upstreamID)
-
-	return upstreamStatus.IsHealthy(maxHeight)
 }
